Apologies for the delay in getting this submitted, but the selector stuff had me scratching my head for a bit.
The underlying problem I ended up with was using minikube locally and its lack of networkpolicy enforcement by default was preventing my testing from running as expected.  But I've verified this on Killcoda and I'm pretty sure it works as desired.

Even if I don't get the job, I appreciate this exercise, it taught me a lot of stuff I should have known for years. --Nick